third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-single-value-in-head-expected.txt

This is a testharness.js-based test.
FAIL Meta color-scheme in head applies. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
Harness: the test ran to completion.
